gpu-worker/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
numpy>=1.26.0
opencv-python-headless>=4.10.0
Pillow>=10.4.0
torch>=2.2.0
torchvision>=0.17.0
python-dotenv>=1.0.0
mediapipe>=0.10.0
